shangchaun

+++ src/main/java/com/volunteer/web/manager/UserInfoBindManager.java
@@ -6,6 +6,6 @@ public interface UserInfoBindManager {
 	/**
 	 *新增绑定用户信息
 	 */
-	Long  saveUserInfoBind(UserInfoBind userInfoBind);
+	Integer  saveUserInfoBind(UserInfoBind userInfoBind);
 
 }
